updating to official repo

+++ platform/iphone/game_center.mm
@@ -5,8 +5,8 @@
 /*                           GODOT ENGINE                                */
 /*                      https://godotengine.org                          */
 /*************************************************************************/
-/* Copyright (c) 2007-2017 Juan Linietsky, Ariel Manzur.                 */
-/* Copyright (c) 2014-2017 Godot Engine contributors (cf. AUTHORS.md)    */
+/* Copyright (c) 2007-2018 Juan Linietsky, Ariel Manzur.                 */
+/* Copyright (c) 2014-2018 Godot Engine contributors (cf. AUTHORS.md)    */
 /*                                                                       */
 /* Permission is hereby granted, free of charge, to any person obtaining */
 /* a copy of this software and associated documentation files (the       */
@@ -27,6 +27,7 @@
 /* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
+
 #ifdef GAME_CENTER_ENABLED
 
 #include "game_center.h"
@@ -109,7 +110,7 @@ void GameCenter::connect() {
 				GameCenter::get_singleton()->authenticated = true;
 			} else {
 				ret["result"] = "error";
-				ret["error_code"] = error.code;
+				ret["error_code"] = (int64_t)error.code;
 				ret["error_description"] = [error.localizedDescription UTF8String];
 				GameCenter::get_singleton()->authenticated = false;
 			};
@@ -138,14 +139,13 @@ Error GameCenter::post_score(Variant p_score) {
 
 	[GKScore reportScores:@[ reporter ]
 			withCompletionHandler:^(NSError *error) {
-
 				Dictionary ret;
 				ret["type"] = "post_score";
 				if (error == nil) {
 					ret["result"] = "ok";
 				} else {
 					ret["result"] = "error";
-					ret["error_code"] = error.code;
+					ret["error_code"] = (int64_t)error.code;
 					ret["error_description"] = [error.localizedDescription UTF8String];
 				};
 
@@ -176,14 +176,13 @@ Error GameCenter::award_achievement(Variant p_params) {
 
 	[GKAchievement reportAchievements:@[ achievement ]
 				withCompletionHandler:^(NSError *error) {
-
 					Dictionary ret;
 					ret["type"] = "award_achievement";
 					if (error == nil) {
 						ret["result"] = "ok";
 					} else {
 						ret["result"] = "error";
-						ret["error_code"] = error.code;
+						ret["error_code"] = (int64_t)error.code;
 					};
 
 					pending_events.push_back(ret);
@@ -195,7 +194,6 @@ Error GameCenter::award_achievement(Variant p_params) {
 void GameCenter::request_achievement_descriptions() {
 
 	[GKAchievementDescription loadAchievementDescriptionsWithCompletionHandler:^(NSArray *descriptions, NSError *error) {
-
 		Dictionary ret;
 		ret["type"] = "achievement_descriptions";
 		if (error == nil) {
@@ -241,7 +239,7 @@ void GameCenter::request_achievement_descriptions() {
 
 		} else {
 			ret["result"] = "error";
-			ret["error_code"] = error.code;
+			ret["error_code"] = (int64_t)error.code;
 		};
 
 		pending_events.push_back(ret);
@@ -251,7 +249,6 @@ void GameCenter::request_achievement_descriptions() {
 void GameCenter::request_achievements() {
 
 	[GKAchievement loadAchievementsWithCompletionHandler:^(NSArray *achievements, NSError *error) {
-
 		Dictionary ret;
 		ret["type"] = "achievements";
 		if (error == nil) {
@@ -273,7 +270,7 @@ void GameCenter::request_achievements() {
 
 		} else {
 			ret["result"] = "error";
-			ret["error_code"] = error.code;
+			ret["error_code"] = (int64_t)error.code;
 		};
 
 		pending_events.push_back(ret);
@@ -289,7 +286,7 @@ void GameCenter::reset_achievements() {
 			ret["result"] = "ok";
 		} else {
 			ret["result"] = "error";
-			ret["error_code"] = error.code;
+			ret["error_code"] = (int64_t)error.code;
 		};
 
 		pending_events.push_back(ret);
@@ -346,7 +343,6 @@ Error GameCenter::request_identity_verification_signature() {
 
 	GKLocalPlayer *player = [GKLocalPlayer localPlayer];
 	[player generateIdentityVerificationSignatureWithCompletionHandler:^(NSURL *publicKeyUrl, NSData *signature, NSData *salt, uint64_t timestamp, NSError *error) {
-
 		Dictionary ret;
 		ret["type"] = "identity_verification_signature";
 		if (error == nil) {
@@ -358,7 +354,7 @@ Error GameCenter::request_identity_verification_signature() {
 			ret["player_id"] = [player.playerID UTF8String];
 		} else {
 			ret["result"] = "error";
-			ret["error_code"] = error.code;
+			ret["error_code"] = (int64_t)error.code;
 			ret["error_description"] = [error.localizedDescription UTF8String];
 		};
